Trees.Trees

Description

Here is a longer description of this module, containing some commentary with some markup.

Documentation

data Btree x

Constructors

Lf x
Btree stands for Binary Tree which consists of a leaf Lf x labelled with a single element x.

Br (Btree x) x (Btree x)
Btree stands for Binary Tree which consists of a branch Br (Btree x) x (Btree x) consisting of two children and a label x.